snowpipe/int128: make the division algorithm clearly from absl
